Move Type.toSchema() to TypeUtil.kt in core/data
