+++ java/arcs/core/host/Utils.kt
@@ -10,13 +10,8 @@
  */
 package arcs.core.host
 
-import arcs.core.data.CollectionType
-import arcs.core.data.EntityType
 import arcs.core.data.Plan
-import arcs.core.data.Schema
-import arcs.core.data.SingletonType
 import arcs.core.host.api.Particle
-import arcs.core.type.Type
 import kotlin.reflect.KClass
 
 /**
@@ -46,26 +41,3 @@ inline fun <reified T : Particle> ((Plan.Particle?) -> T).toRegistration(): Part
     val construct: suspend (Plan.Particle?) -> T = { this.invoke(it) }
     return T::class.toParticleIdentifier() to construct
 }
-/**
- * If this Type represents a [SingletonType], [CollectionType], or [EntityType], return the
- * [Schema] used by the underlying [Entity] that this type represents.
- */
-fun Type.toSchema(): Schema {
-    when (this) {
-        is SingletonType<*> -> if (this.containedType is EntityType) {
-            return (this.containedType as EntityType).entitySchema
-        }
-        is CollectionType<*> -> if (this.collectionType is EntityType) {
-            return (this.collectionType as EntityType).entitySchema
-        }
-        is EntityType -> return this.entitySchema
-        else -> Unit
-    }
-    throw IllegalArgumentException("Can't get entitySchema of unknown type $this")
-}
-
-/**
-* If this Type represents a [SingletonType], [CollectionType], or [EntityType], return the
-* [Schema.hash] used by the underlying [Entity] that this type represents.
-*/
-fun Type.toSchemaHash(): String = this.toSchema().hash
